Revamp the token bridge module
This patch updates the token bridge module and changes how it works in
several ways

1) No more support for sending ETH directly, gas management caused the
   majority of the complexity in this module and also frustration for
   end users, who expect to get their total value into the router. In
   order to address this we will be sending ETH to routers to pay the
   fees. So no more ETH management is required.
2) Module refactor splitting what was previously one file into three to
   help with readability and make room for an upcoming Alhtea blockchain
   module

This will require an update to the dashboard as extensive changes to the
display states have also been made. But we're now envisoning a deposit
situation where the user does not really look at the router. So detailed
display isn't as high on our list of priorities.
